upd dto variable names to match with backend response,

+++ mobile/app/src/main/java/com/example/mobile/remote/dtos/Report/ReportResponse.kt
@@ -2,5 +2,5 @@ package com.example.mobile.remote.dtos.Report
 
 data class ReportResponse (
     val date: String,
-    val amount:Int
+    val amount:Float
 )
